# docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: PostFlow API
  version: 0.2.0
  description: v2 API for multi-network, multi-account publishing via account_id.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
  - apiKeyAuth: []
  - basicAuth: []
paths:
...
  /accounts/static:
    post:
      summary: Create or update a static account for LinkedIn, Facebook, or Instagram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStaticAccountRequest'
      responses:
        '201':
          description: Account created/updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/Error'
...
  /login:
    get:
      summary: Render the local owner login page
      responses:
        '200':
          description: HTML login page

    post:
      summary: Create a local owner web session
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                return_to:
                  type: string
      responses:
        '303':
          description: Redirect to the requested HTML page after login
        '400':
          $ref: '#/components/responses/Error'
...
  /oauth/register:
    post:
      summary: Register a public OAuth client for the local MCP server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_name:
                  type: string
                redirect_uris:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Registered OAuth client
        '400':
          $ref: '#/components/responses/Error'
  /authorize:
    get:
      summary: Authorize a registered OAuth client using the current owner web session
      description: If the owner is not logged in yet, PostFlow redirects to `/login` first and then resumes the OAuth flow.
      parameters:
        - in: query
          name: response_type
          required: true
          schema:
            type: string
            enum: [code]
        - in: query
          name: client_id
          required: true
          schema:
            type: string
        - in: query
          name: redirect_uri
          required: true
          schema:
            type: string
        - in: query
          name: code_challenge
          required: true
          schema:
            type: string
        - in: query
          name: code_challenge_method
          required: true
          schema:
            type: string
            enum: [S256]
        - in: query
          name: scope
          schema:
            type: string
        - in: query
          name: state
          schema:
            type: string
      responses:
        '302':
          description: Redirect back to the client redirect URI with an authorization code
        '303':
          description: Redirect to local login if no owner session exists
        '400':
          $ref: '#/components/responses/Error'
  /token:
    post:
      summary: Exchange or refresh OAuth tokens for the local MCP server
      description: Returns bearer access and refresh tokens for the local MCP server. Remote MCP clients use these tokens for authenticated `tools/call` requests.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                grant_type:
                  type: string
                client_id:
                  type: string
                redirect_uri:
                  type: string
                code:
                  type: string
                code_verifier:
                  type: string
                refresh_token:
                  type: string
      responses:
        '200':
          description: OAuth bearer token pair
        '400':
          $ref: '#/components/responses/Error'
  /media:
    post:
      summary: Upload a media file
      description: Uploads are capped at 512 MiB.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                kind:
                  type: string
                  enum: [video, image]
                  default: video
                file:
                  type: string
                  format: binary
                tags:
                  type: string
                  description: Comma-separated editorial tags for later media search.
      responses:
        '201':
          description: Media created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        '400':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
    get:
      summary: List uploaded media
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - in: query
          name: tag
          schema:
            type: string
          description: Optional editorial tag filter.
      responses:
        '200':
          description: Media list
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaListItem'
  /media/{id}:
    delete:
      summary: Delete media if not in use
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Media deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                  media:
                    $ref: '#/components/schemas/MediaListItem'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
  /campaigns:
    get:
      summary: List editorial campaigns
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/CampaignStatus'
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Campaign list
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Campaign'
    post:
      summary: Create an editorial campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignInput'
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          $ref: '#/components/responses/Error'
...
  /editorial/backlog:
    get:
      summary: List editorial backlog items
      parameters:
        - in: query
          name: campaign_id
          schema:
            type: string
        - in: query
          name: platform
          schema:
            $ref: '#/components/schemas/Platform'
        - in: query
          name: editorial_status
          schema:
            $ref: '#/components/schemas/EditorialStatus'
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Editorial backlog
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/EditorialBacklogItem'
  /posts:
    post:
      summary: Create a draft or scheduled post
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Post'
                  - $ref: '#/components/schemas/CreatePostBatchResponse'
        '200':
          description: Idempotent replay (existing post)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Post'
                  - $ref: '#/components/schemas/CreatePostBatchResponse'
        '400':
          $ref: '#/components/responses/Error'
  /posts/validate:
    post:
      summary: Validate a post without persisting it (dry-run)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '200':
          description: Validation succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  normalized:
                    type: object
                    properties:
                      account_id:
                        type: string
                      platform:
                        $ref: '#/components/schemas/Platform'
                      text:
                        type: string
                      scheduled_at:
                        type: string
                        format: date-time
                      media_ids:
                        type: array
                        items:
                          type: string
                      segments:
                        type: array
                        items:
                          $ref: '#/components/schemas/ThreadSegmentInput'
                      max_attempts:
                        type: integer
                  warnings:
                    type: array
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/Error'
...
  /drafts:
    get:
      summary: List draft posts (no scheduled date)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
      responses:
        '200':
          description: Draft list
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  drafts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
  /schedule:
    get:
      summary: List scheduled publications by range
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: view
          schema:
            type: string
            enum: [publications, posts]
            default: publications
      responses:
        '200':
          description: Schedule list
          content:
            application/json:
              schema:
                type: object
                properties:
                  from:
                    type: string
                    format: date-time
                  to:
                    type: string
                    format: date-time
                  count:
                    type: integer
                  items:
                    oneOf:
                      - type: array
                        items:
                          $ref: '#/components/schemas/Publication'
                      - type: array
                        items:
                          $ref: '#/components/schemas/Post'
  /dlq:
    get:
      summary: List dead-letter queue entries
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: DLQ item list
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeadLetter'
  /dlq/requeue:
    post:
      summary: Requeue multiple DLQ entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Bulk requeue result
          content:
            application/json:
              schema:
                type: object
                properties:
                  selected:
                    type: integer
                  success:
                    type: integer
                  failed:
                    type: integer
        '400':
          $ref: '#/components/responses/Error'
  /dlq/delete:
    post:
      summary: Delete multiple DLQ entries and their failed posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Bulk delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  selected:
                    type: integer
                  success:
                    type: integer
                  failed:
                    type: integer
        '400':
          $ref: '#/components/responses/Error'
...
  /settings/timezone:
    post:
      summary: Set UI timezone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [timezone]
              properties:
                timezone:
                  type: string
                  example: Europe/Madrid
      responses:
        '200':
          description: Timezone updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  timezone:
                    type: string
        '400':
          $ref: '#/components/responses/Error'
...
  /content-plans/preview:
    post:
      summary: Preview a multi-brand content plan
      description: Validates the date range, cadence, accounts, and brand profiles and returns the number of shared ideas, network variants, and generated images without persisting anything.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentPlanInput'
      responses:
        '200':
          description: Plan volume estimate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentPlanPreview'
        '400':
          $ref: '#/components/responses/Error'
  /content-plans:
    get:
      summary: List content plans
      responses:
        '200':
          description: Content plans with their current progress
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContentPlan'
    post:
      summary: Create an editable content plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentPlanInput'
      responses:
        '201':
          description: Draft content plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentPlan'
        '400':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    basicAuth:
      type: http
      scheme: basic
...
  responses:
    Error:
      description: Validation or business error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
  schemas:
...
    Account:
      type: object
      properties:
        id:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
        account_kind:
          $ref: '#/components/schemas/AccountKind'
        display_name:
          type: string
        external_account_id:
          type: string
        auth_method:
          type: string
          enum: [static, oauth]
        status:
          type: string
          enum: [connected, disconnected, error]
        last_error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateStaticAccountRequest:
      type: object
      required: [platform, external_account_id, credentials]
      description: Static accounts are supported for linkedin, facebook, and instagram. X accounts must connect via OAuth.
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        account_kind:
          $ref: '#/components/schemas/AccountKind'
        display_name:
          type: string
        external_account_id:
          type: string
        credentials:
          type: object
          additionalProperties: true
...